Security
Completed the user authentication process, implemented access control for dashboard section and refactored some code
