+++ config/packages/security.yaml
@@ -16,6 +16,15 @@ security:
         main:
             lazy: true
             provider: app_user_provider
+            form_login:
+                login_path: daema_login
+                check_path: daema_login
+                enable_csrf: true
+                default_target_path: daema_dashboard # Redirect path after successful login
+            logout:
+                path: daema_logout
+                # where to redirect after logout
+                target: daema_home
 
             # activate different ways to authenticate
             # https://symfony.com/doc/current/security.html#the-firewall
@@ -27,7 +36,7 @@ security:
     # Note: Only the *first* access control that matches will be used
     access_control:
         # - { path: ^/admin, roles: ROLE_ADMIN }
-        # - { path: ^/profile, roles: ROLE_USER }
+        - { path: ^/daema/dashboard, roles: ROLE_USER }
 
 when@test:
     security:
