+++ templates/daema/login.html.twig
@@ -1,39 +1,64 @@
+{# templates/security/login.html.twig #}
+
 {% extends 'base.html.twig' %}
 
-{% block title %}Login{% endblock %}
+{% block title %}Log in!{% endblock %}
 
-{% block stylesheets %}<link rel="stylesheet" href="{{ asset('styles/daema/login.css')}}">{% endblock %}
+{% block stylesheets %}
+<link rel="stylesheet" href="{{ asset('styles/daema/login.css') }}">
+{% endblock %}
 
 {% block body %}
 <div class="bg-gradient-to-r from-rose-100 to-teal-100 overflow-hidden">
-
     <!-- ============================================================== -->
     <!-- Start Page Content here -->
     <!-- ============================================================== -->
 
     <div class="h-screen w-screen flex justify-center items-center">
-
         <div class="2xl:w-1/4 lg:w-1/3 md:w-1/2 w-full">
             <div class="card overflow-hidden sm:rounded-md rounded-none bg-slate-50 shadow">
                 <div class="p-6">
+                    {% if error %}
+                        <div class="alert alert-danger">{{ error.messageKey|trans(error.messageData, 'security') }}</div>
+                    {% endif %}
 
-                    <div class="mb-4">
-                        <label class="block text-sm font-medium text-gray-600 mb-2" for="LoggingUsername">Username</label>
-                        <input id="LoggingUsername" class="form-input" type="username" placeholder="Enter your username" required>
-                    </div>
+                    {% if app.user %}
+                        <div class="mb-3">
+                            You are logged in as {{ app.user.userIdentifier }}, <a href="{{ path('daema_logout') }}">Logout</a>
+                        </div>
+                    {% endif %}
 
-                    <div class="mb-4">
-                        <label class="block text-sm font-medium text-gray-600 mb-2" for="loggingPassword">Password</label>
-                        <input id="loggingPassword" class="form-input" type="password" placeholder="Enter your password" required>
-                    </div>
+                    <h1 class="text-xl font-bold mb-4">Log In</h1>
 
-                    <div class="flex justify-center mb-6">
-                        <button type"submit" class="btn w-auto text-white bg-blue-600 hover:bg-blue-700"> Log In </button>
+                    <form method="post">
+                        <div class="mb-4">
+                            <label class="block text-sm font-medium text-gray-600 mb-2" for="username">Username</label>
+                            <input type="text" value="{{ last_username }}" name="_username" id="username" class="form-input" autocomplete="username" required autofocus>
+                        </div>
+
+                        <div class="mb-4">
+                            <label class="block text-sm font-medium text-gray-600 mb-2" for="password">Password</label>
+                            <input type="password" name="_password" id="password" class="form-input" autocomplete="current-password" required>
+                        </div>
+
+                        <input type="hidden" name="_csrf_token" value="{{ csrf_token('authenticate') }}">
+
+                        <div class="checkbox mb-4">
+                            <label>
+                                <input type="checkbox" name="_remember_me"> Remember me
+                            </label>
+                        </div>
+
+                        <div class="flex justify-center mb-6">
+                            <button type="submit" class="btn w-auto text-white bg-blue-600 hover:bg-blue-700"> Log In </button>
+                        </div>
+                    </form>
+                    <div class="text-center mt-4">
+                        <p class="text-gray-600">Don't have an account? <a href="{{ path('daema_register') }}" class="text-blue-600 hover:text-blue-800 font-bold">Register</a></p>
                     </div>
                 </div>
             </div>
         </div>
     </div>
 </div>
-
 {% endblock %}
