+++ templates/daema/register.html.twig
@@ -39,6 +39,10 @@
                     </div>
                     
                     {{ form_end(registrationForm) }}
+
+                    <div class="text-center mt-4">
+                        <p class="text-gray-600">Already have an account? <a href="{{ path('daema_login') }}" class="text-blue-600 hover:text-blue-800 font-bold">Log In</a></p>
+                    </div>
                 </div>
             </div>
         </div>
